Tidying code up a bit
