+++ src/components/Footer.tsx
@@ -15,7 +15,7 @@ const socials = [
 
 export const Footer = () => {
     return (
-        <Container component='footer' sx={{ position: 'fixed', width: '100%', bottom: 0 }}>
+        <Container component='footer' sx={{ position: 'fixed', bottom: 0 }}>
             <Box>  
                 {socials.map( ( social ) => (   
                     <Link href={social.url}>
@@ -25,7 +25,7 @@ export const Footer = () => {
                     </Link>
                 ))}
             </Box>
-            <Typography>Made by Harry Allom</Typography>
+            <Typography color="text.secondary">Made by Harry Allom</Typography>
         </Container>
     );
 }
